♻️ refactor(config): rename config to changelogconfig
- rename Config struct to ChangeLogConfig for clarity
- update all references to Config to use ChangeLogConfig
